Update hypervolume_estimate_probability.R
add check for NA values in `points`. Previously, `na.omit(points)` was used but this failed if predicting to raster as the number of returned probabilities would not equal the number of cell in the raster. Now, if NA values are found, an attribute is created and attached to the returned probabilities showing where the NA values occurred. These cells are then ignored when assigning probabilities to rasters for the output of `hypervolume_project`.